sharper/requirements.txt
Flask
Flask-Assets
Flask-SQLAlchemy
Flask-Cache
Flask-Script
Flask-WTF
Pillow
hiredis
redis
pyqqweibo
python-dateutil
requests
simplejson
Fabric
BeautifulSoup
blinker
lunardate
